Set up API handlers
- Use a module structure for the handlers that encapsulates the API
  models with their endpoints.
- Add API docs URLs to the metadata response.
