+++ src/hoverdrive/handlers/external/__init__.py
@@ -1,0 +1,3 @@
+from .endpoints import router as external_router
+
+__all__ = ["external_router"]
